chore: fmt, clippy, and the two follow-ups from the warnings pass (#313)
* style: cargo fmt both workspaces

Pure formatting, no semantic change. These 16 files predate the warnings
pass and were left alone there so that PR's diff stayed reviewable; taking
them separately here keeps this commit trivially skimmable and the clippy
work that follows honest.

Note the two workspaces are on different editions — the app root is 2024,
bevy-tutti is 2021 — so they must be formatted through their own manifests.
A bare `rustfmt --edition 2021` over changed files reorders app-workspace
imports the wrong way.

* fix: clear clippy across both workspaces

The app workspace was already clean; these are the engine's ~74.

Most were mechanical (`clippy --fix`): `then_some` over a closure, `Copy`
types cloned, `Into` impls that should be `From`, redundant casts, manual
`div_ceil`, derivable `impl`s, test modules sitting mid-file.

Three needed judgment, and two of them clippy was **wrong** about:

- `!(sample_rate.get() > 0.0)` in `tutti-analysis` (`geometry.rs`, `yin.rs`)
  is not awkward style, it is the NaN guard. Every comparison against NaN is
  false, so the suggested `x <= 0.0` *accepts* a NaN rate and lets it reach
  the FFT geometry — verified, not assumed. Allowed with the reason on the
  line, since the idiom is correct but reads as a typo.

- `needless_range_loop` in `fft.rs` and `downmix_unit.rs`: in both, the index
  is a coordinate rather than a cursor. `fft.rs` walks two slices in opposite
  directions (`buffer[size - i]` against `bins[i]`) to mirror a spectrum;
  `downmix_unit.rs` uses `(c, i)` as a (channel, frame) pair that `set_f32`
  takes as two arguments. An iterator rewrite obscures both.

- `vst3.rs`'s `match` → `if let`: the discarded `Err` arm carried the comment
  explaining *why* a refused reactivation is swallowed. Restructured so that
  reasoning stays attached rather than stranded above an `if let Ok`.

Two real (if small) findings in `vst-tutti`: `payload.as_ptr() as *const u8
as *mut u8` was a redundant cast wrapping a load-bearing const→mut one, and
six `as *const *const _` casts on FFI call arguments were pure no-ops.

Engine suite: 3657 passed, 0 failed. Both workspaces clean under default,
`--all-targets` and `--all-features`.

* fix(vst2): make DispatcherProc unsafe, as its signature always implied

`DispatcherProc` takes four raw pointers across an FFI boundary and hands
them to arbitrary third-party plugin code, but was declared a *safe*
`extern "C" fn`. Calling it therefore needed no `unsafe`, so the obligations
a caller actually carries — a live `AEffect` the pointer belongs to, an
`opcode`-appropriate `ptr`, a `value` the plugin will interpret — were
nowhere stated and could not be reviewed.

This was not theoretical. Earlier in this branch's history, three `unsafe`
blocks around `dispatch` calls in `lib.rs`'s tests were *removed* as
redundant, and they were redundant, precisely because the type understated
its contract. With the type corrected, the compiler asks for all three back.
That round trip is the argument for the change.

Six call sites, each now carrying a SAFETY comment: one real host→plugin
dispatch in `host.rs`, four in tests, and one in the reference plugin's
dispatcher shim, where the obligations are the host's and were met on the
way in.

Scope is deliberately just this one type. `ProcessProc`, `SetParameterProc`,
`GetParameterProc` and `HostCallbackProc` have the same defect, but they are
called from ~10 files across `tutti-vst2-host` and belong in their own
change rather than riding along with a lint pass.

vst-tutti + tutti-vst2-host + tutti-plugin(vst2): 202 passed, 0 failed.

* test(vst2): cover the loader, which had no tests at all

`loaders::vst2` was the one format shim with zero tests. That surfaced in
the previous commit as a warning rather than as a gap: `PLUGIN_LOAD_LOCK`
had to be gated away from `vst2`, because nothing in a vst2-only build ever
took it — even though VST2's non-thread-safe `LOAD_POINTER` is the reason
the lock exists.

Wired the reference plugin on the same terms as its CLAP sibling:
`tutti-vst2-test-plugin` becomes a dev-dependency so cargo builds the cdylib
inside this same `cargo test`, and `build.rs` resolves where it landed via
the shared `tutti-fixture-resolve`. No new mechanism — `emit_candidates` /
`resolve_or_panic` is the existing pattern, and the probe's behaviour is
steered per test through `TUTTI_VST2_PROBE_*`.

Nine tests, aimed at the invariants the loader's own comments assert:

- one bus per side, with the *width* checked too — `single_bus` takes a
  layout where a count would also compile
- an asymmetric 1-in/4-out case, which is what catches a transposition; the
  stereo case alone passes against a loader that hard-codes 2
- `TRANSPORT` unconditional vs `EDITOR` measured, asserted together against
  an editor-less plugin, so wiring both to one source fails
- `tail: Unknown` rather than `None` — the loader's longest comment, and
  audible: `None` tells a bounce to add no decay and truncates every VST2
  reverb
- empty speaker topology, since VST2 cannot be asked
- a missing file failing at `Opening` rather than panicking

Mutation-tested, per the repo's policy — tail→None, inputs/outputs swapped,
TRANSPORT wired to the editor flag, EDITOR forced on: each fails the suite.

One test found my own premise wrong rather than the code. An early draft
asserted `MIDI_OUTPUTS=0` clears `MIDI_OUT`; it does not, because
`tutti-vst2-host` resolves `can_do(SendMidiEvent)` first and only falls back
to pin counts on `Maybe` — an explicit yes outranks an inference, correctly.
The test now pins the seam this loader actually owns (two host fields → two
bits) and says why the inference path belongs to `tutti-vst2-host`'s
integration suite, which can drive the `dlopen` switches.

Engine 3666 passed (+9), app 1381 passed, 0 failed. All 16 loader feature
combinations still clean.

+++ crates/dsp/tutti-analysis/src/pitch.rs
@@ -118,7 +118,7 @@ impl PitchDetector {
         let frequency = Hz((self.sample_rate.get() / refined_period) as f32);
         // Clamped rather than wrapped raw: `Confidence` is a 0..=1 reading, and
         // this is the one place the aperiodicity inversion could leave the range.
-        let confidence = Confidence::new_clamped(1.0 - aperiodicity as f32);
+        let confidence = Confidence::new_clamped(1.0 - aperiodicity);
 
         PitchResult {
             frequency,
